fix: add EPERM check and fix setgroups signature per review
Based on code review feedback:
- Add EPERM check for CAP_SETGID (requires euid == 0 since capabilities not implemented)
- Change setgroups size parameter from usize to i32 to match kernel syscall signature
- Add test for negative size in setgroups
- Add documentation about thread-sharing limitation

Security: setgroups now properly requires root privileges to prevent
privilege escalation attacks.

+++ litebox_common_linux/src/lib.rs
@@ -2226,7 +2226,7 @@ pub enum SyscallRequest<Platform: litebox::platform::RawPointerProvider> {
         list: Option<Platform::RawMutPointer<u32>>,
     },
     Setgroups {
-        size: usize,
+        size: i32,
         list: Option<Platform::RawConstPointer<u32>>,
     },
     Sysinfo {
